backend/requirements.txt
flask
torch
torchvision
Pillow
llama-cpp-python